Add functionality to save user profiles
Add a 'profile' field to the UserInfo struct. Modify the
'exchange_authorization_code' method to retrieve the profile information
using the profile provider and set the profile field of the UserInfo
struct when the biome-profile feature is enabled.

Modify the 'make_callback_route' method to take a UserProfileStore as an
argument and save the profile information stored in the profile field of
the UserInfo struct if the biome-profile feature is enabled. If the
given profile already exists in the profile store it will be updated
otherwise it will be added as a new profile.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>